refactoring ground and tossing controllers
